# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de suivi de poids et séances d'entraînement",
        "title": "API App-Sport",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Quentin Lachery",
            "email": "quentin.lkss@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8020",
    "basePath": "/",
    "paths": {
...
        "/seances": {
            "get": {
                "description": "Retourne toutes les séances enregistrées",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Séances"
                ],
                "summary": "Liste des séances",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Seance"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Ajoute une nouvelle séance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Séances"
                ],
                "summary": "Créer une séance",
                "parameters": [
                    {
                        "description": "Séance à enregistrer",
                        "name": "seance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SeanceInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Seance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/seances/by-date": {
            "get": {
                "description": "Récupère toutes les séances pour une date spécifique",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Séances"
                ],
                "summary": "Séances par date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date au format YYYY-MM-DD",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Seance"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/seances/today": {
            "get": {
                "description": "Récupère toutes les séances prévues pour la date du jour",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Séances"
                ],
                "summary": "Séances du jour",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Seance"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/seances/{id}": {
            "get": {
                "description": "Retourne une séance par son ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Séances"
                ],
                "summary": "Détail d'une séance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la séance",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Seance"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Met à jour une séance existante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Séances"
                ],
                "summary": "Modifier une séance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la séance",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Champs à modifier",
                        "name": "seance",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SeanceInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Seance"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Supprime une séance par ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Séances"
                ],
                "summary": "Supprimer une séance",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la séance",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Exercice": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "groupe_musculaire": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nom": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.ExerciceRealise": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "exercice": {
                    "$ref": "#/definitions/models.Exercice"
                },
                "exercice_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "ordre": {
                    "type": "integer"
                },
                "poids": {
                    "type": "number"
                },
                "reps": {
                    "type": "integer"
                },
                "rpe": {
                    "type": "number"
                },
                "seance_id": {
                    "type": "integer"
                },
                "series": {
                    "type": "integer"
                }
            }
        },
...
        "models.Seance": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "exercices_realises": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ExerciceRealise"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.SeanceInput": {
            "type": "object",
            "properties": {
                "comment": {
                    "type": "string"
                },
                "date": {
                    "description": "Format YYYY-MM-DD",
                    "type": "string"
                },
                "type": {
                    "description": "Muscu, natation, etc.",
                    "type": "string"
                }
            }
        }
    }
}